release
开启负载网络:wifi -s ssid password maxpeer
获取客户机mac: wifi -ls
关闭负载网络: wifi-q
