+++ wifi/wifi/x64/Debug/wifi.Build.CppClean.log
@@ -1,0 +1,14 @@
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\vc140.pdb
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\vc140.idb
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\main.obj
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\main_netsh.obj
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\test.obj
+e:\github\lastingguy\attendancemanagement\wifi\x64\debug\wifi.ilk
+e:\github\lastingguy\attendancemanagement\wifi\x64\debug\wifi.exe
+e:\github\lastingguy\attendancemanagement\wifi\x64\debug\wifi.pdb
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\wifi.tlog\cl.command.1.tlog
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\wifi.tlog\cl.read.1.tlog
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\wifi.tlog\cl.write.1.tlog
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\wifi.tlog\link.command.1.tlog
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\wifi.tlog\link.read.1.tlog
+e:\github\lastingguy\attendancemanagement\wifi\wifi\x64\debug\wifi.tlog\link.write.1.tlog
